// src/types/FormInput.ts
export interface FormInput {
  character: boolean;
  location: boolean;
  episodes: boolean;
  char_name: string;
  char_status: string;
  char_species: string;
  char_type: string;
  char_gender: string;
  loc_name: string;
  loc_type: string;
  loc_dimension: string;
  ep_name: string;
  ep_code: string;
}